Move input components somewhere sane
