+++ src/components/inputs/CurrencyInput.vue
@@ -1,10 +1,10 @@
 <script setup lang="ts">
 import type { PropType } from "vue";
 import type { Dinero } from "dinero.js";
-import ActionButton from "./buttons/ActionButton.vue";
+import ActionButton from "../buttons/ActionButton.vue";
 import TextField from "./TextField.vue";
 import { dinero } from "dinero.js";
-import { intlFormat } from "../transformers";
+import { intlFormat } from "../../transformers";
 import { ref, computed, toRefs, watch } from "vue";
 import { USD } from "@dinero.js/currencies";
