+++ src/components/inputs/DateTimeInput.vue
@@ -1,4 +1,5 @@
 <script lang="ts">
+// Loads before `setup` runs
 const currentDate = new Date();
 currentDate.setSeconds(0, 0);
 </script>
@@ -6,10 +7,10 @@ currentDate.setSeconds(0, 0);
 <script setup lang="ts">
 import type { PropType } from "vue";
 import "vue3-date-time-picker/dist/main.css";
-import ActionButton from "./buttons/ActionButton.vue";
+import ActionButton from "../buttons/ActionButton.vue";
 import DatePicker from "vue3-date-time-picker";
 import { computed, toRefs } from "vue";
-import { useUiStore } from "../store/uiStore";
+import { useUiStore } from "../../store/uiStore";
 
 const emit = defineEmits(["update:modelValue"]);
