+++ src/components/settings/ChangePasswordForm.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import ActionButton from "../buttons/ActionButton.vue";
-import TextField from "../TextField.vue";
+import TextField from "../inputs/TextField.vue";
 import { computed, ref } from "vue";
 import { useAuthStore } from "../../store/authStore";
 import { useToast } from "vue-toastification";
