+++ src/components/transactions/TransactionEdit.vue
@@ -4,14 +4,14 @@ import type { Location, LocationRecordParams } from "../../model/Location";
 import type { PropType } from "vue";
 import type { TransactionRecordParams } from "../../model/Transaction";
 import ActionButton from "../buttons/ActionButton.vue";
-import Checkbox from "../Checkbox.vue";
+import Checkbox from "../inputs/Checkbox.vue";
 import CheckmarkIcon from "../../icons/Checkmark.vue";
 import ConfirmDestroyTransaction from "./ConfirmDestroyTransaction.vue";
-import CurrencyInput from "../CurrencyInput.vue";
-import DateTimeInput from "../DateTimeInput.vue";
+import CurrencyInput from "../inputs/CurrencyInput.vue";
+import DateTimeInput from "../inputs/DateTimeInput.vue";
 import LocationField from "../locations/LocationField.vue";
-import TextAreaField from "../TextAreaField.vue";
-import TextField from "../TextField.vue";
+import TextAreaField from "../inputs/TextAreaField.vue";
+import TextField from "../inputs/TextField.vue";
 import TrashIcon from "../../icons/Trash.vue";
 import { dinero, isNegative, isZero, toSnapshot } from "dinero.js";
 import { ref, computed, toRefs, onMounted } from "vue";
