+++ src/pages/Login.vue
@@ -2,7 +2,7 @@
 import ActionButton from "../components/buttons/ActionButton.vue";
 import ErrorNotice from "../components/ErrorNotice.vue";
 import Footer from "../Footer.vue";
-import TextField from "../components/TextField.vue";
+import TextField from "../components/inputs/TextField.vue";
 import { ref, computed, watch, onMounted, nextTick } from "vue";
 import { useRoute, useRouter } from "vue-router";
 import { useAuthStore } from "../store/authStore";
